# Conflicts:
#	Scoring/ScoringService/src/main/java/Kafka/KafkaReceiver/processor/ScoreServiceProcessor.java

+++ ChangeDetectorServices/change-detector-service/src/main/java/edu/miu/changedetectorservice/service/StandardDeviationService.java
@@ -1,8 +1,9 @@
 package edu.miu.changedetectorservice.service;
 
+import org.springframework.stereotype.Service;
 
 public interface StandardDeviationService {
 
-    void findDataChange(Long x);
+    void findDataChange(Double x);
 }
