+++ ChangeDetectorServices/change-detector-service/src/main/java/edu/miu/changedetectorservice/storage/InMemoryStorage.java
@@ -4,7 +4,7 @@ import java.util.LinkedList;
 import java.util.Queue;
 
 public class InMemoryStorage {
-    public static final Queue<Long> VALUE_HOLDING = new LinkedList<>();
+    public static final Queue<Double> VALUE_HOLDING = new LinkedList<>();
     public static Double PREVIOUS_SD = 0D;
 
     private InMemoryStorage(){}
